examples/cpp/data-collector/gui/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyside6 = "^6.8.0"
numpy = "^2.1.2"
pyzmq = "^26.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
